Updating `DebuggerSupportsLLVMTarget` to call lldb::SBDebugger::GetBuildConfiguration, which is static and renamed the skip macro to match the python name.
