+++ lldb/unittests/DAP/TestBase.h
@@ -10,43 +10,17 @@
 #include "DAPLog.h"
 #include "Handler/RequestHandler.h"
 #include "Handler/ResponseHandler.h"
-#include "Protocol/ProtocolBase.h"
 #include "TestingSupport/Host/JSONTransportTestUtilities.h"
 #include "TestingSupport/SubsystemRAII.h"
 #include "Transport.h"
 #include "lldb/Host/FileSystem.h"
 #include "lldb/Host/HostInfo.h"
 #include "lldb/Host/MainLoop.h"
-#include "lldb/Host/MainLoopBase.h"
 #include "llvm/ADT/StringRef.h"
 #include "llvm/Support/FileSystem.h"
-#include "llvm/Support/JSON.h"
-#include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include <memory>
 #include <optional>
-
-/// Helpers for gtest printing.
-namespace lldb_dap::protocol {
-
-inline void PrintTo(const Request &req, std::ostream *os) {
-  *os << llvm::formatv("{0}", toJSON(req)).str();
-}
-
-inline void PrintTo(const Response &resp, std::ostream *os) {
-  *os << llvm::formatv("{0}", toJSON(resp)).str();
-}
-
-inline void PrintTo(const Event &evt, std::ostream *os) {
-  *os << llvm::formatv("{0}", toJSON(evt)).str();
-}
-
-inline void PrintTo(const Message &message, std::ostream *os) {
-  return std::visit([os](auto &&message) { return PrintTo(message, os); },
-                    message);
-}
-
-} // namespace lldb_dap::protocol
 
 namespace lldb_dap_tests {
 
@@ -74,38 +48,11 @@ protected:
   void Run();
 };
 
-/// A matcher for a DAP event.
-template <typename EventMatcher, typename BodyMatcher>
-inline testing::Matcher<const lldb_dap::protocol::Event &>
-IsEvent(const EventMatcher &event_matcher, const BodyMatcher &body_matcher) {
-  return testing::AllOf(
-      testing::Field(&lldb_dap::protocol::Event::event, event_matcher),
-      testing::Field(&lldb_dap::protocol::Event::body, body_matcher));
-}
-
-template <typename EventMatcher>
-inline testing::Matcher<const lldb_dap::protocol::Event &>
-IsEvent(const EventMatcher &event_matcher) {
-  return testing::AllOf(
-      testing::Field(&lldb_dap::protocol::Event::event, event_matcher),
-      testing::Field(&lldb_dap::protocol::Event::body, std::nullopt));
-}
-
-/// Matches an "output" event.
-inline auto Output(llvm::StringRef o, llvm::StringRef cat = "console") {
-  return IsEvent("output",
-                 testing::Optional(llvm::json::Value(
-                     llvm::json::Object{{"category", cat}, {"output", o}})));
-}
-
 /// A base class for tests that interact with a `lldb_dap::DAP` instance.
 class DAPTestBase : public TransportBase {
 protected:
   std::optional<llvm::sys::fs::TempFile> core;
   std::optional<llvm::sys::fs::TempFile> binary;
-
-  static constexpr llvm::StringLiteral k_linux_binary = "linux-x86_64.out.yaml";
-  static constexpr llvm::StringLiteral k_linux_core = "linux-x86_64.core.yaml";
 
   static void SetUpTestSuite();
   static void TearDownTestSuite();
@@ -113,7 +60,7 @@ protected:
   void TearDown() override;
 
   void CreateDebugger();
-  void LoadCore();
+  void LoadCore(llvm::StringRef binary_path, llvm::StringRef core_path);
 };
 
 } // namespace lldb_dap_tests
